! apply custom setters for update
