+++ lib/lhs/concerns/item/update.rb
@@ -25,7 +25,7 @@ class LHS::Item < LHS::Proxy
 
     def update!(params, options = {}, partial_update = false)
       options ||= {}
-      partial_data = LHS::Data.new(params, _data.parent, record)
+      partial_data = record.new(LHS::Data.new(params, _data.parent, record))
       _data.merge_raw!(partial_data)
       data_sent = partial_update ? partial_data : _data
       url = href || record.find_endpoint(id: id).compile(id: id)
