+++ spec/item/update_spec.rb
@@ -48,6 +48,35 @@ describe LHS::Item do
       expect(item.name).to eq 'Andrea'
       expect(item.likes).not_to eq 'Banana'
     end
+
+    context 'with custom setters' do
+      before do
+        class Booking < LHS::Record
+          endpoint 'http://bookings/bookings'
+
+          def appointments=(appointments)
+            super(
+              appointments.map { |appointment| appointment[:id] }
+            )
+          end
+        end
+      end
+
+      let(:item) do
+        Booking.new(id: 'abc')
+      end
+
+      it 'updates data using custom setters before send to backend' do
+        stub_request(:post, "http://bookings/bookings")
+          .with(body: {
+            id: 'abc',
+            appointments: [1, 2, 3]
+          }.to_json)
+          .to_return(status: 200, body: "", headers: {})
+        item.update(appointments: [{ id: 1 }, { id: 2 }, { id: 3 }])
+        expect(item.appointments.to_a).to eq([1, 2, 3])
+      end
+    end
   end
 
   context 'update!' do
